Share tests for prepared statements
