+++ sqlite3/test/wasm/prepared_statement_test.dart
@@ -1,0 +1,6 @@
+import '../common/prepared_statement.dart';
+import 'utils.dart';
+
+void main() {
+  testPreparedStatements(loadSqlite3);
+}
